refactor: update frontend to new proto
