+++ frontend/src/pages/Register.tsx
@@ -27,7 +27,7 @@ import {
   RegisterRequest,
   RegisterResponse,
   UserCredentials,
-} from "../proto/user-bff-service";
+} from "../proto/user-service";
 import useFixedToast from "../utils/hooks/useFixedToast";
 
 function Register() {
